+++ backend/packages/caching/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cat-factory/caching
 
+## 0.6.4
+
+### Patch Changes
+
+- @cat-factory/kernel@0.104.4
+
+## 0.6.3
+
+### Patch Changes
+
+- @cat-factory/kernel@0.104.3
+
 ## 0.6.2
 
 ### Patch Changes