+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cat-factory/provider-s3
 
+## 0.2.109
+
+### Patch Changes
+
+- @cat-factory/kernel@0.104.4
+
+## 0.2.108
+
+### Patch Changes
+
+- @cat-factory/kernel@0.104.3
+
 ## 0.2.107
 
 ### Patch Changes