+++ backend/packages/spend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @cat-factory/spend
 
+## 0.11.10
+
+### Patch Changes
+
+- Updated dependencies [4a3e536]
+  - @cat-factory/contracts@0.117.0
+  - @cat-factory/kernel@0.104.4
+
+## 0.11.9
+
+### Patch Changes
+
+- Updated dependencies [18a9cb5]
+  - @cat-factory/contracts@0.116.1
+  - @cat-factory/kernel@0.104.3
+
 ## 0.11.8
 
 ### Patch Changes